+++ igstkImageSliceViewer.cxx
@@ -256,9 +256,10 @@ ImageSliceViewer
     return;
     }
 
-  double spacing[3];
-  double origin[3];
+  float spacing[3];
+  float origin[3];
   int   dimensions[3];
+
 
   image->GetSpacing(spacing);
   image->GetOrigin(origin);
